+++ Meraki.Api/Data/DeviceProvisioningStatusNetwork.cs
@@ -1,0 +1,13 @@
+﻿namespace Meraki.Api.Data;
+
+[DataContract]
+[ApiAccessReadOnlyClass]
+public class DeviceProvisioningStatusNetwork
+{
+	/// <summary>
+	/// ID for the network containing the device
+	/// </summary>
+	[DataMember(Name = "id")]
+	public string? Id { get; set; }
+
+}