+++ Meraki.Api/Data/ProductsUpdateRequest.cs
@@ -10,41 +10,53 @@ public class ProductsUpdateRequest
 	/// Appliance
 	/// </summary>
 	[DataMember(Name = "appliance")]
-	public Appliance Appliance { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? Appliance { get; set; }
 
 	/// <summary>
 	/// Camera
 	/// </summary>
 	[DataMember(Name = "camera")]
-	public Appliance Camera { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? Camera { get; set; }
 
 	/// <summary>
 	/// Cellular gateway
 	/// </summary>
 	[DataMember(Name = "cellularGateway")]
-	public Appliance CellularGateway { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? CellularGateway { get; set; }
 
 	/// <summary>
 	/// Environmental
 	/// </summary>
 	[DataMember(Name = "environmental")]
-	public Appliance Environmental { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? Environmental { get; set; }
 
 	/// <summary>
 	/// Switch
 	/// </summary>
 	[DataMember(Name = "switch")]
-	public SwitchUpdate Switch { get; set; } = new();
+	public SwitchUpdate? Switch { get; set; }
 
 	/// <summary>
 	/// VMX Host
 	/// </summary>
 	[DataMember(Name = "vmxHost")]
-	public Appliance VmxHost { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? VmxHost { get; set; }
 
 	/// <summary>
 	/// Wireless
 	/// </summary>
 	[DataMember(Name = "wireless")]
-	public Appliance Wireless { get; set; } = new();
+	public NetworkFirmwareUpdateRequestProduct? Wireless { get; set; }
+
+	/// <summary>
+	/// CloudGateway
+	/// </summary>
+	[DataMember(Name = "cloudGateway")]
+	public NetworkFirmwareUpdateRequestProduct? CloudGateway { get; set; }
+
+	/// <summary>
+	/// SwitchCatalyst
+	/// </summary>
+	[DataMember(Name = "switchCatalyst")]
+	public NetworkFirmwareUpdateRequestProduct? SwitchCatalyst { get; set; }
 }