+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ﻿# Changelog
 
+## 1.34.1
+- Updated for v1.34 API changes
+
+## 1.33.22
+- Updated ClientUsageHistory Sent and Received to be longs instead of ints.
+
+## 1.33.6
+- Added uplinkClientSampling to Network Switch Settings
+
+## 1.33.1
+- Updated for v1.33 API changes
+
 ## 1.32.7
 - Added observed value "open" for EncryptionMode on SSID